Refactor operations into Operator classes so adding new operations is easier from the outside.  Add each operation into it's own operator class.

+++ src/main/groovy/gratum/etl/LoadStatistic.groovy
@@ -52,7 +52,7 @@ class LoadStatistic {
     }
 
     public Integer getRejections() {
-        return rejectionsByCategory.inject(0) { Integer sum, RejectionCategory cat, Integer count -> sum + count }
+        return (Integer)rejectionsByCategory.inject(0) { Integer sum, RejectionCategory cat, Integer count -> sum + count }
     }
 
     public void reject(RejectionCategory category) {
